% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ccdr-main-R.R
\name{ccdr.gridR}
\alias{ccdr.gridR}
\title{ccdr.gridR}
\usage{
ccdr.gridR(X, betas, lambdas, nlam = 50, gamma = 2, eps = 1e-04,
  rlam = 1e-04, maxIters, alpha = 2, verbose = FALSE, as.mat = TRUE)
}
\arguments{
\item{X}{data matrix}

\item{betas}{OPTIONAL beta matrix / DEFAULT = zeroes}

\item{lambdas}{OPTIONAL lambda grid / DEFAULT = derived from data}

\item{nlam}{OPTIONAL number of lambdas / DEFAULT = 50}

\item{gamma}{OPTIONAL gamma parameter for MCP / DEFAULT = 2.0}

\item{eps}{OPTIONAL error threshold / DEFAULT = 1e-4}

\item{rlam}{OPTIONAL ratio between min.lam : max.lam / DEFAULT = 1e-4}

\item{maxIters}{OPTIONAL maximum number of iterations / DEFAULT = 2 * max(10, sqrt(pp))}

\item{alpha}{OPTIONAL stopping criterion / DEFAULT = 2.0}

\item{verbose}{OPTIONAL flag for printing out progress / DEFAULT = FALSE}

\item{as.mat}{OPTIONAL convert results to R Matrix class or not?}
}
\description{
ccdr.gridR
}

