% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ccdr.R
\docType{package}
\name{ccdr}
\alias{ccdr}
\alias{ccdr-package}
\title{ccdr: Structure learning for Bayesian networks using the CCDr algorithm.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

