% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ccdr-main-R.R
\name{ccdr.run}
\alias{ccdr.run}
\title{ccdr.run}
\usage{
ccdr.run(data, betas, lambdas, lambdas.length = 20, gamma = 2,
  error.tol = 1e-04, lambdas.ratio = 0.01, max.iters, alpha = 10,
  verbose = FALSE)
}
\description{
Placeholder for what will eventually be the main function exported from this package
}

