% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ccdr-main-R.R
\name{ccdr.singleR}
\alias{ccdr.singleR}
\title{ccdr.singleR}
\usage{
ccdr.singleR(c, pp, nn, betas, lambda, gamma = 2, eps = 1e-04,
  rlam = 1e-04, maxIters, alpha = 2, verbose = FALSE)
}
\arguments{
\item{c}{correlation vector}

\item{pp}{}

\item{nn}{}

\item{betas}{OPTIONAL beta matrix / DEFAULT = zeroes}

\item{gamma}{OPTIONAL gamma parameter for MCP / DEFAULT = 2.0}

\item{eps}{OPTIONAL error threshold / DEFAULT = 1e-4}

\item{rlam}{OPTIONAL ratio between min.lam : max.lam / DEFAULT = 1e-4}

\item{maxIters}{OPTIONAL maximum number of iterations / DEFAULT = 2 * max(10, sqrt(pp))}

\item{alpha}{OPTIONAL stopping criterion / DEFAULT = 2.0}

\item{verbose}{OPTIONAL flag for printing out progress / DEFAULT = FALSE}
}
\description{
ccdr.singleR
}

