% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdrAlgorithm.R
\docType{package}
\name{ccdrAlgorithm}
\alias{ccdrAlgorithm}
\alias{ccdrAlgorithm-package}
\title{ccdrAlgorithm: CCDr Algorithm for Learning Sparse Gaussian Bayesian Networks}
\description{
Implements the CCDr structure learning algorithm as described in
\href{http://arxiv.org/abs/1401.0852}{Aragam and Zhou (2015)}.
}
\details{
The CCDr algorithm uses sparse regularization (L1 or MCP) to produce
a solution path of DAG estimates along a pre-determined grid of hyperparameters.
This package implements a single function, \code{\link{ccdr.run}} that
runs the main algorithm, and uses \code{\link[sparsebnUtils]{sparsebnUtils}}
for the underlying data structures and methods.
}
