% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-ccdrFit.R
\docType{S3class}
\name{ccdrFit-class}
\alias{ccdrFit-class}
\title{ccdrFit class}
\description{
Convenience wrapper class for output of CCDr algorithm: Represents a single DAG estimate in the solution path.
}
\section{Slots}{

\describe{
\item{\code{sbm}}{Should eventually be a graph object (probably).}
\item{\code{lambda}}{(numeric) Value of lambda for this estimate.}
\item{\code{nedge}}{(integer) Number of edges in this estiamte.}
\item{\code{pp}}{(integer) Number of nodes.}
\item{\code{nn}}{(integer) Number of observations this estimate was based on.}
\item{\code{time}}{(numeric) Time in seconds to generate this estimate.}
}
}

\section{Methods}{

\code{\link{get.adjacency.matrix}}, \code{\link{get.sigmas}}
}

