% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-ccdrPath.R
\docType{class}
\name{ccdrPath-class}
\alias{ccdrPath-class}
\title{ccdrPath class}
\description{
Convenience wrapper class for output of CCDr algorithm: Represents the entire solution path
of the CCDr algorithm. Its components are of type \code{\link{ccdrFit-class}}. Also inherits
from \code{\link{list}}.
}
\details{
Each value of lambda in the (discrete) solution path corresponds to a single DAG estimate, which
is of the form (Phi, Rho) (see \href{http://arxiv.org/abs/1401.0852}{Aragam and Zhou (2015), JMLR} for details).
Internally, this estimate is represented by a \code{\link{ccdrFit-class}} object. The full solution
path is then represented as a \code{\link{list}} of \code{\link{ccdrFit-class}} objects: This class is essentially a wrapper for this list.
}
\section{Methods}{

\code{\link{get.adjacency.matrix}}, \code{\link{lambda.grid}},
\code{\link{num.nodes}}, \code{\link{num.edges}}, \code{\link{num.samples}}
}

