% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-ccdrFit.R
\name{edge.list.ccdrFit}
\alias{edge.list.ccdrFit}
\title{edge.list.ccdrFit}
\usage{
\method{edge.list}{ccdrFit}(cf)
}
\value{
\code{matrix}
}
\description{
Extracts the edge list associated with the DAG of a \code{\link{ccdrFit-class}} object.
}

