% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-edgeList.R
\docType{S3class}
\name{edgeList-class}
\alias{edgeList-class}
\title{edgeList class}
\description{
Convenience wrapper class for a (column-major) edge list. Each component of the list
corresponds to a node, and each component is a list whose components are the parents
of this node in the graph. Only intended for internal use.
}
\section{Slots}{

\describe{
\item{\code{edges}}{Edge list representation of a graph sorted by children.}
}
}

\section{Methods}{

\code{\link{get.adjacency.matrix}}
}

