% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-ccdrFit.R
\name{get.adjacency.matrix.ccdrFit}
\alias{get.adjacency.matrix.ccdrFit}
\title{get.adjacency.matrix.ccdrFit}
\usage{
\method{get.adjacency.matrix}{ccdrFit}(cf)
}
\value{
\code{matrix}
}
\description{
Extracts the adjacency matrix from a \code{\link{ccdrFit-class}} object.
}

