% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-ccdrPath.R
\name{get.adjacency.matrix.ccdrPath}
\alias{get.adjacency.matrix.ccdrPath}
\title{get.adjacency.matrix.ccdrPath}
\usage{
\method{get.adjacency.matrix}{ccdrPath}(cp)
}
\description{
get.adjacency.matrix.ccdrPath
}

