% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-SparseBlockMatrixR.R, R/s3-ccdrFit.R, R/s3-ccdrPath.R, R/s3-edgeList.R, R/s3-generics.R
\name{get.adjacency.matrix.SparseBlockMatrixR}
\alias{get.adjacency.matrix}
\alias{get.adjacency.matrix.SparseBlockMatrixR}
\alias{get.adjacency.matrix.ccdrFit}
\alias{get.adjacency.matrix.ccdrPath}
\alias{get.adjacency.matrix.edgeList}
\title{get.adjacency.matrix}
\usage{
\method{get.adjacency.matrix}{SparseBlockMatrixR}(sbm)

\method{get.adjacency.matrix}{ccdrFit}(cf)

\method{get.adjacency.matrix}{ccdrPath}(cp)

\method{get.adjacency.matrix}{edgeList}(edgeL)

get.adjacency.matrix(x)
}
\value{
\code{matrix}
}
\description{
Extracts the adjacency matrix of the associated graph object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SparseBlockMatrixR}: Convert internal SparseBlockMatrixR representation to an adjacency matrix

\item \code{ccdrFit}: Retrieves \code{edges} slot and converts to an adjacency matrix

\item \code{ccdrPath}: Retrieves all \code{edges} slots in the solution path, converts to an adjacency matrix, and returns as a list

\item \code{edgeList}: Convert internal edgeList representation to an adjacency matrix
}}

