% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-ccdrFit.R
\name{get.sigmas.ccdrFit}
\alias{get.sigmas.ccdrFit}
\title{get.sigmas.ccdrFit}
\usage{
\method{get.sigmas}{ccdrFit}(cf)
}
\value{
\code{numeric}
}
\description{
Extracts the vector of inverse variances from a \code{\link{ccdrFit-class}} object.
}

