% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-SparseBlockMatrixR.R, R/s3-edgeList.R, R/s3-generics.R
\name{is.zero.SparseBlockMatrixR}
\alias{is.zero}
\alias{is.zero.SparseBlockMatrixR}
\alias{is.zero.edgeList}
\title{is.zero}
\usage{
\method{is.zero}{SparseBlockMatrixR}(x)

\method{is.zero}{edgeList}(edgeL)

is.zero(x)
}
\description{
Determines whether or not the object is the same as the null or zero object from its class.
}
\section{Methods (by class)}{
\itemize{
\item \code{SparseBlockMatrixR}: 

\item \code{edgeList}: 
}}

