% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-SparseBlockMatrixR.R, R/s3-ccdrFit.R, R/s3-ccdrPath.R, R/s3-edgeList.R, R/s3-generics.R
\name{num.edges.SparseBlockMatrixR}
\alias{num.edges}
\alias{num.edges.SparseBlockMatrixR}
\alias{num.edges.ccdrFit}
\alias{num.edges.ccdrPath}
\alias{num.edges.edgeList}
\title{num.edges}
\usage{
\method{num.edges}{SparseBlockMatrixR}(sbm)

\method{num.edges}{ccdrFit}(cf)

\method{num.edges}{ccdrPath}(cp)

\method{num.edges}{edgeList}(edgeL)

num.edges(x)
}
\description{
Extracts the number of edges of the associated graph object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SparseBlockMatrixR}: 

\item \code{ccdrFit}: 

\item \code{ccdrPath}: 

\item \code{edgeList}: 
}}

