% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-SparseBlockMatrixR.R, R/s3-ccdrFit.R, R/s3-ccdrPath.R, R/s3-edgeList.R, R/s3-generics.R
\name{num.nodes.SparseBlockMatrixR}
\alias{num.nodes}
\alias{num.nodes.SparseBlockMatrixR}
\alias{num.nodes.ccdrFit}
\alias{num.nodes.ccdrPath}
\alias{num.nodes.edgeList}
\title{num.nodes}
\usage{
\method{num.nodes}{SparseBlockMatrixR}(sbm)

\method{num.nodes}{ccdrFit}(cf)

\method{num.nodes}{ccdrPath}(cp)

\method{num.nodes}{edgeList}(edgeL)

num.nodes(x)
}
\description{
Extracts the number of nodes of the associated graph object.
}
\section{Methods (by class)}{
\itemize{
\item \code{SparseBlockMatrixR}: 

\item \code{ccdrFit}: 

\item \code{ccdrPath}: 

\item \code{edgeList}: 
}}

