% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-ccdrFit.R, R/s3-ccdrPath.R, R/s3-generics.R
\name{num.samples.ccdrFit}
\alias{num.samples}
\alias{num.samples.ccdrFit}
\alias{num.samples.ccdrPath}
\title{num.samples}
\usage{
\method{num.samples}{ccdrFit}(cf)

\method{num.samples}{ccdrPath}(cp)

num.samples(x)
}
\description{
Extracts the number of samples of the associated object.
}
\section{Methods (by class)}{
\itemize{
\item \code{ccdrFit}: 

\item \code{ccdrPath}: 
}}

