% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-ccdrPath.R
\name{print.ccdrPath}
\alias{print.ccdrPath}
\title{print.ccdrPath}
\usage{
\method{print}{ccdrPath}(cp, verbose = FALSE)
}
\arguments{
\item{verbose}{If \code{TRUE}, then each estimate in the solution path is printed separately. Do not use for
       large graphs or large solution paths. (default = \code{FALSE})}
}
\description{
Prints the contents of a \code{\link{ccdrPath-class}} object neatly.
}

