% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-generics.R
\docType{import}
\name{reexports}
\alias{get.adjacency.matrix}
\alias{is.zero}
\alias{num.edges}
\alias{num.nodes}
\alias{reIndexC}
\alias{reIndexR}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{sparsebnUtils}{\code{\link[sparsebnUtils]{reIndexC}}, \code{\link[sparsebnUtils]{reIndexR}}, \code{\link[sparsebnUtils]{get.adjacency.matrix}}, \code{\link[sparsebnUtils]{num.nodes}}, \code{\link[sparsebnUtils]{num.edges}}, \code{\link[sparsebnUtils]{is.zero}}}
}}
\keyword{internal}

