% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3-ccdrFit.R
\docType{S3class}
\name{ccdrFit-class}
\alias{ccdrFit-class}
\title{ccdrFit class}
\description{
Main class for representing DAG estimates: Represents a single DAG estimate in the solution path.
Generally speaking, these estimates should be wrapped up in a \code{\link{ccdrPath-class}} object, but
can be handled separately if desired (be careful!).
}
\section{Slots}{

\describe{
\item{\code{edge}}{(edgeList) Edge list of estimated DAG (see \code{\link{edgeList-class}}).}
\item{\code{lambda}}{(numeric) Value of lambda for this estimate.}
\item{\code{nedge}}{(integer) Number of edges in this estimate.}
\item{\code{pp}}{(integer) Number of nodes.}
\item{\code{nn}}{(integer) Number of observations this estimate was based on.}
\item{\code{time}}{(numeric) Time in seconds to generate this estimate.}
}
}

\section{Methods}{

\code{\link{get.adjacency.matrix}}
\code{\link{num.nodes}}, \code{\link{num.edges}}, \code{\link{num.samples}}
}

