% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdrAlgorithm-mvn.R
\name{generate_mvn_data}
\alias{generate_mvn_data}
\title{Generate data from a DAG}
\usage{
generate_mvn_data(graph, params, n = 1, ivn = NULL, ivn.rand = TRUE)
}
\arguments{
\item{graph}{DAG in \code{\link{edgeList}} format.}

\item{params}{Vector of parameters. Last p elements correspond to variances (p = number of nodes in \code{graph}), initial elements correspond to edge weights.}

\item{n}{Number of samples to draw.}

\item{ivn}{List of interventions (see \code{\link[sparsebnUtils]{sparsebnData}}). Must be a \code{list} with exactly \code{n} components.}

\item{ivn.rand}{If \code{TRUE}, random N(0,1) values will be drawn for each intervention. Otherwise, these values need to supplied manually in \code{ivn}.}
}
\description{
Given a Gaussian DAG, generate data from the underlying distribution.
Equivalently, generate data from a multivariate normal distribution given
one of its SEM. Can generate both observational and intervention data.
}
\details{
If \code{ivn = NULL}, then \code{n} observational samples are drawn. For each
component of \code{ivn} that is not \code{NULL}, interventional samples will
be drawn with the values of each node specified in the component.
}
